requirement.txt
django 
djangorestframework 
django-corsheaders 
python-decouple 
drf_spectacular